Increase default `commit.retry.num-retries` for Iceberg tables (#23)
Because Lake Loader is a horizontally scalable application, it is
possible to get CommitFailed exceptions when multiple loaders try to
commit at the same time. Here we increase the number of retries from 4
to 20. This is a table property not a runtime property, so it only
affects newly created tables.