Move the content of deltaIceberg module to core module (#24)
With the removal of hudi and biglake variants from the project in the previous
commit, only deltaIceberg package is left for providing table format setup errors.

In this commit, we move the content of the package to core module and remove it since
we don't need it as separate package anymore.